refactor: check for pip on uninstall

+++ scripts/release/linux-macOS_remove.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
-pip uninstall cyberdrop-dl
-pip uninstall cyberdrop-dl-patched
+if command -v pip >/dev/null 2>&1; then
+    pip uninstall cyberdrop-dl
+    pip uninstall cyberdrop-dl-patched
+fi
 uv tool uninstall cyberdrop-dl-patched
 uv cache clean
